+++ src/Sulu/Bundle/SearchBundle/CHANGELOG.md
@@ -2,8 +2,9 @@ CHANGELOG
 =========
 
 0.2
----
+-----------
 
+- "webspaceKey" is now always indexed
 - Deindex unpublished nodes
 - Map Url, Description and Title fields from structure template to index metadata
 - Event listener for setting the portal prefix on the document URL